refactor: auto-split test_validation_comprehensive.py
